!35 优化POSTEMB 和  数据处理 * add  format_style file * move compute size to preprocess * 修改onnx图 input5-12为int32 * 修改input5-12为int，修改postemb为仅支持half消除warning

+++ src/python/modify_ERNIE.py
@@ -135,7 +135,7 @@ fuser.fuse()
 if ENABLE_POSTEMBEDDING_PLUGIN:
     graph.inputs=graph.inputs[:5]
     graph.inputs[4].shape=(-1,8)
-    graph.inputs[4].dtype=np.float32
+    graph.inputs[4].dtype=np.int32
     graph.inputs[4].name="read_file_0.tmp_6-13"
     
 if DEBUG:
